+++ gem/__init__.py
@@ -74,6 +74,7 @@ class Path:
     User        = path_join(xdg_config_home, "gem")
     Data        = path_join(xdg_data_home, "gem")
     Apps        = path_join(xdg_data_home, "applications")
+    Lock        = path_join(xdg_data_home, ".lock")
     Roms        = path_join(xdg_data_home, "gem", "roms")
     Logs        = path_join(xdg_data_home, "gem", "logs")
     Notes       = path_join(xdg_data_home, "gem", "notes")