+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 0.6.1 - Dr. Funfrock
 
 * Add windows size memorization
+* Add a lock to avoid multi GEM instances
+* Fix block preferences when some games are launched
 * Fix manage preferences window as Dialog when open from GEM window
 * Fix preferences window crash when use drag and drop in main window while preferences is open
 * Fix specify bash for install.sh